Refactor GraphQL helpers
